[CAF] Adds basic support for a set of coarray collectives
This adds some basic support for a set of coarray collectives that have been
proposed for inclusion into the Fortran standard:

   co_sum, co_product, co_maxval, co_minval

Front-end support also was added for:

   co_all, co_any, co_count, co_findloc, co_maxloc, co_inloc

A few general CAF bug fixes are also in this commit:

* In some cases the front-end would reuse a dope vector of a coarray type for a
  non-coarray variable, which confused the subsequent back-end lowering phase.
  This commit prevents such a thing from happening.

* The compiler didn't handle coscalars as arguments in CAF intrinsic calls,
  specifically this_image, image_index, lcobound, and ucobound. This commit
  adds support for this in the front-end. It required generation of a dope
  vector for non-allocatable coscalars and some other changes.

Also, there were some minor adjustments to the functions names that are
created by the compiler for consistency.

Contributor: Deepak, Siddhartha

git-svn-id: https://svn.open64.net/svnroot/open64/branches/OpenUH@3944 ead328b1-e903-0410-a19d-eacee4b031f6

+++ osprey/libcaf/armci/armci_comm_layer.c
@@ -32,8 +32,9 @@
 #include <assert.h>
 #include <unistd.h>
 #include <math.h>
-
+#include "caf_rtl.h"
 #include "armci_comm_layer.h"
+#include "trace.h"
 #include "util.h"
 
 /* common_slot is a node in the shared memory link-list that keeps track
@@ -108,13 +109,13 @@ static void wait_on_all_pending_puts();
  * Inline functions
  */
 /* must call comm_init() first */
-inline unsigned long comm_get_proc_id()
+inline size_t comm_get_proc_id()
 {
     return my_proc;
 }
 
 /* must call comm_init() first */
-inline unsigned long comm_get_num_procs()
+inline size_t comm_get_num_procs()
 {
     return num_procs;
 }
